test(bridge): keep the !windows cleanup test beside its tagged implementation
cleanup.go is //go:build !windows, but bridge_cleanup_test.go carried no tag
and called cmdlineHasExactArg from it. On Windows the whole internal/bridge
test binary therefore failed to compile, and because `go vet` type-checks test
files, `go vet ./...` failed with it:

    internal\bridge\bridge_cleanup_test.go:204:5: undefined: cmdlineHasExactArg

Move the one unix-only test into cleanup_unix_test.go under the same build
constraint as the code it exercises. The rest of bridge_cleanup_test.go is
already portable and stays where it is.

No production code changes. `go vet ./...` now passes on windows/amd64, and
internal/bridge builds and runs there.
