+++ internal/bridge/bridge_cleanup_test.go
@@ -195,20 +195,6 @@ func TestCleanup_CloakTermBeforeKill(t *testing.T) {
 	}
 }
 
-func TestCmdlineHasExactUserDataDirArgAvoidsPrefixCollision(t *testing.T) {
-	args := [][]byte{
-		[]byte("/Applications/Chrome"),
-		[]byte("--user-data-dir=/var/lib/pt/profile10"),
-		[]byte("--remote-debugging-port=9222"),
-	}
-	if cmdlineHasExactArg(args, "--user-data-dir=/var/lib/pt/profile1") {
-		t.Fatal("profile1 matched profile10")
-	}
-	if !cmdlineHasExactArg(args, "--user-data-dir=/var/lib/pt/profile10") {
-		t.Fatal("profile10 should match exact user-data-dir arg")
-	}
-}
-
 func TestCleanup_RemoteCDP_SkipsKillLogic(t *testing.T) {
 	withShortenedGrace(t, 20*time.Millisecond, 20*time.Millisecond)
 	termCount, killCount := withMockedProcessLookup(t, func() []int { return []int{1} })
